% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSatelliteSeries.R
\name{addHarmonicFits}
\alias{addHarmonicFits}
\title{Fit harmonics recursivelys to satellite time series}
\usage{
addHarmonicFits(ts, iterations, origin, dependent, ..., dt_col = DATE,
  omega = 1)
}
\arguments{
\item{ts}{Data.frame containing columns date, dependent variable, and any
grouping variables.}

\item{iterations}{Number of recursive harmonics to fit}

\item{origin}{reference time; format should match df date column}

\item{dependent}{column name for dependent variable}

\item{...}{additional grouping variables; unquoted}

\item{dt_col}{unquoted column name for date column. Default is DATE}

\item{omega}{default is 1}
}
\description{
Runs a 2-term harmonic regression for the specified number of iterations
}
\examples{
# apply harmonics by pointID
library(dplyr)
library(lubridate)
library(stringr)
library(broom)
df_fitted <- df_satelliteSeries \%>\%
  addHarmonicFits(
    iterations = 4,
    origin = ymd("2018-01-01"),
    dependent = "GCVI",
    pointID
 )
}
\keyword{harmonics}
