% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy_generic.R
\name{calcStats}
\alias{calcStats}
\title{Generic accuracy assessment function}
\usage{
calcStats(x_pred1, y_obs1)
}
\arguments{
\item{x_pred}{column name for predictions}

\item{y_obs}{column name reference/validation observations}

\item{df}{data frame object}
}
\description{
Calculates a suite of agreement statistics for two sets of paired observations.
}
\examples{
# get statistics for all observations
agreement <- calcStats('x_pred','y_obs', df)

# get statistics for groups of observations
library(dplyr)
agreementByGroup <- df \%>\%
  group_by(group) \%>\%
  group_map(~calcStats(.$x_pred,.$y_obs))
}
\keyword{accuracy}
\keyword{metrics}
