% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentinelCalibration.R
\name{getCalibrationCoeffs}
\alias{getCalibrationCoeffs}
\title{Get Sentinel TOA to Sentinel SR Calibration}
\usage{
getCalibrationCoeffs(sentineldf, toaColName, srColName)
}
\arguments{
\item{sentineldf}{Data.frame containing paired Sentinel TOA and SR values for a band or index}

\item{toaColName}{column name for Sentinel TOA values}

\item{srColName}{column name for Senintel SR values}
}
\description{
Calculates a linear equation between Sentinel SR and TOA. Returns the
intercept and slope as a data frame, along with the r^2
}
\examples{
# get linear coefficients
coefficients <- getCalibrationCoeffs(sentinelData_df, 'Sentinel_TOA', 'Sentinel_SR')
}
\keyword{Calibration}
\keyword{Sentinel}
