% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSatelliteSeries.R
\name{getPeakGCVIs_fromFile}
\alias{getPeakGCVIs_fromFile}
\title{Extract the maximum index value and DOY for each point from raw csv files}
\usage{
getPeakGCVIs_fromFile(df_file, datasource, maskClouds = FALSE,
  w1_doy1 = 152, w1_doy2 = 245)
}
\arguments{
\item{df_file}{Data.frame containing a veg index of interest}

\item{datasource}{"Landsat" or "Sentinel"}

\item{maskClouds}{should cloud qa be applied? defaults to FALSE}

\item{w1_doy1}{first day of first window; default is June 1}

\item{w1_doy2}{last day of first window; default is Sept 1}
}
\description{
Reduces a time series of satellite observations for each point to a maximum
value for the specified window. Requires readr, lubridate, and dplyr.
Default window is June - August.
Initial setup is for
exports from Jill's GEE script '02.1x_satelliteSampler_forScymOffline'
}
\examples{
# to apply to a list of files:
landsatFiles <- list.files(gDrive, pattern=landsatBaseName, full.names = TRUE)
# check file sizes and drop empty exports
landsatFilesb <- landsatFiles[sapply(landsatFiles, file.size) > 2]
# load/get window GCVIs - default windows
landsat_gcvi <- purrr::map_df(landsatFilesb, getPeakGCVIs_fromFile,
                               datasource = "Landsat", maskClouds = TRUE)
}
\keyword{SCYM}
\keyword{peak}
\keyword{preparation,}
\keyword{vi}
