% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSatelliteSeries.R
\name{getRecursiveCoeffs}
\alias{getRecursiveCoeffs}
\title{Get final recursive harmonics model coefficients}
\usage{
getRecursiveCoeffs(ts, iterations, origin, dependent0, ...,
  dt_col = DATE, omega = 1)
}
\arguments{
\item{ts}{Data.frame containing columns date, dependent variable, and any
grouping variables.}

\item{iterations}{Number of recursive harmonics to fit}

\item{origin}{reference time; format should match df date column}

\item{...}{additional grouping variables; unquoted}

\item{dt_col}{unquoted column name for date column. Default is DATE}

\item{omega}{default is 1}

\item{dependent}{column name for dependent variable}
}
\description{
Runs a 2-term harmonic regression for the specified number of iterations.
Modifies getHarmonicFits to return the coefficients instead of fitted values
}
\examples{
# apply harmonics by pointID
library(dplyr)
library(lubridate)
library(stringr)
library(broom)

coefficients <- df_satelliteSeries \%>\%
  getRecursiveCoeffs(iterations = numIterations,
                     origin = ymd("2018-01-01"),
                     dependent = "GCVI",
                      omega = 1.5, pointID)
}
\keyword{harmonics}
