% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSatelliteSeries.R
\name{getWindowGCVIs_fromFile}
\alias{getWindowGCVIs_fromFile}
\title{Extract the maximum index value by crop window for each point from raw csv files}
\usage{
getWindowGCVIs_fromFile(df_file, datasource, maskClouds = FALSE,
  w1_doy1 = 165, w1_doy2 = 200, w2_doy1 = 201, w2_doy2 = 240)
}
\arguments{
\item{df_file}{Data.frame containing a veg index of interest}

\item{datasource}{"Landsat" or "Sentinel"}

\item{maskClouds}{should cloud qa be applied? defaults to FALSE}

\item{w1_doy1}{first day of first window; default is US maize}

\item{w1_doy2}{last day of first window; default is US maize}

\item{w2_doy1}{first day of second window; default is US maize}

\item{w2_doy2}{last day of second window; default is US maize}
}
\description{
Reduces a time series of satellite observations for each point to a maximum
value for the specified windows. Requires readr, lubridate, and dplyr.
Initial setup is for
exports from Jill's GEE script '02.1x_satelliteSampler_forScymOffline'
}
\examples{
# to apply to a list of files:
landsatFiles <- list.files(gDrive, pattern=landsatBaseName, full.names = TRUE)
# check file sizes and drop empty exports
landsatFilesb <- landsatFiles[sapply(landsatFiles, file.size) > 2]
# load/get window GCVIs - default windows
landsat_gcvi <- purrr::map_df(landsatFilesb, getWindowGCVIs_fromFile,
                               datasource = "Landsat", maskClouds = TRUE)
}
\keyword{GCVI;}
\keyword{SCYM}
\keyword{US}
\keyword{default}
\keyword{is}
\keyword{maize}
\keyword{preparation,}
