% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSatelliteSeries.R
\name{predictHarmonics}
\alias{predictHarmonics}
\title{Predict value from harmonics}
\usage{
predictHarmonics(doy, constant, a1, a2, b1, b2, omega = 1.5)
}
\arguments{
\item{doy}{doy for prediction}

\item{omega}{default is 1.5}
}
\description{
Predict value from harmonics
}
\keyword{harmonics,}
\keyword{prediction}
